arena: back buf with [u32] to eliminate byte-shuffle in read/write_word
All heap and stack accesses were already 4-byte aligned; switching the
backing buffer from &[u8] to &[u32] lets read_word/write_word compile to
a single LDR/STR instead of a byte-load-and-combine sequence.  Internal
byte offsets are preserved (heap_top, stack_bot, Value::offset() etc.
all remain byte values); the word index is derived with a single >> 2 at
the call site.  Unsafe byte views are used only for the Bytes ops
(alloc_bytes, bytes_data, bytes_concat) and the debug heap_data /
stack_data accessors.

All callers updated: Vm::new, test buffers, and the baremetal example.

+++ crates/shamrocq/src/vm.rs
@@ -129,7 +129,7 @@ pub struct Vm<'buf> {
 }
 
 impl<'buf> Vm<'buf> {
-    pub fn new(buf: &'buf mut [u8]) -> Self {
+    pub fn new(buf: &'buf mut [u32]) -> Self {
         Vm {
             arena: Arena::new(buf),
             globals: [Value::ctor(0, 0); 64],
